use GNU Make to build dataset
refactor scripts
update pip requirements
