+++ scripts/requirements.pip
@@ -1,2 +1,4 @@
 lxml==2.3
 csvkit
+PyICU # https://github.com/Homebrew/legacy-homebrew/issues/34170#issuecomment-63192387
+pandas
